feat: Enhance Koodo Sync functionality and user experience
- Added new translations for Koodo Sync feature in Chinese.
- Refactored backup dialog component to streamline imports.
- Updated header component to prompt users to enable Koodo Sync for improved synchronization experience.
- Modified account settings to handle cloud synchronization upon login.
- Integrated cloud sync functionality into sync settings and login components.
- Implemented new utility functions for handling cloud sync tokens and configurations.
- Improved error handling and user feedback during sync operations.

+++ src/containers/settings/accountSetting/component.tsx
@@ -8,6 +8,7 @@ import {
   formatTimestamp,
   getServerRegion,
   getWebsiteUrl,
+  handleCloudSync,
   handleContextMenu,
   openInBrowser,
   reloadManager,
@@ -189,8 +190,16 @@ class AccountSetting extends React.Component<
       });
       this.props.handleFetchAuthed();
       this.props.handleFetchLoginOptionList();
-      ConfigService.removeItem("defaultSyncOption");
-      ConfigService.removeItem("dataSourceList");
+      let result = await handleCloudSync();
+      if (result) {
+        this.props.cloudSyncFunc();
+      } else {
+        ConfigService.removeItem("defaultSyncOption");
+        ConfigService.removeItem("dataSourceList");
+      }
+      toast.success(this.props.t("Login successful"), {
+        id: "login",
+      });
       this.props.handleFetchDataSourceList();
       this.props.handleFetchDefaultSyncOption();
       this.props.handleFetchUserInfo();
@@ -601,10 +610,15 @@ class AccountSetting extends React.Component<
                     "We have two server regions(Global and China). To change the server region, you need to log out first. Do you want to log out now?"
                   );
                   if (result) {
-                    ConfigService.setItem(
-                      "serverRegion",
-                      getServerRegion() === "china" ? "global" : "china"
-                    );
+                    let newRegion =
+                      getServerRegion() === "china" ? "global" : "china";
+                    ConfigService.setItem("serverRegion", newRegion);
+                    this.setState({
+                      serverRegion: newRegion,
+                    });
+                    resetReaderRequest();
+                    resetUserRequest();
+                    resetThirdpartyRequest();
                     await this.handleLogout();
                   }
                 }}
